소수점 연산의 정확도를 위해 float 대신 double형을 사용 모든 기능을 테스트하는 코드 작성

+++ src/main/java/lotto/Model/Result.java
@@ -35,6 +35,6 @@ public class Result {
                 System.out.println(message);
             }
         }
-        System.out.printf("총 수익률은 %,.1f%%입니다.", (float)totalPrice / purchasePrice * 100);
+        System.out.printf("총 수익률은 %,.1f%%입니다.", (double)totalPrice / purchasePrice * 100);
     }
 }
